🐛 Fix DownloadCounter making ∞ GitHub requests
A bug in the parsing of the Next Link caused this service to make infinite requests for pages of the GitHub API
The bug was in the ghutils library, but was already fixed
We depended on ghreleases, which depended on ghrepos, which depended on ghutils. However, ghreleases called out to ghrepos directly, so we don't need this dependency - it doesn't add anything. ghrepos has ghutils fixed to a ~ version number, so this means we can update to the latest version of ghutils.

I have:

- Switched out ghreleases for ghrepos
- Updated the code with the single line from ghreleases
- Ensured that yarn.lock is set to 3.2.4 - the fixed version of ghutils